+++ test/GasKing.t.sol
@@ -5,10 +5,10 @@ import { PRBTest } from "@prb/test/src/PRBTest.sol";
 import { console2 } from "forge-std/src/console2.sol";
 import { StdCheats } from "forge-std/src/StdCheats.sol";
 
-import {GasKingGame} from "../src/GasKing.sol";
+import { GasKingGame } from "../src/GasKing.sol";
 
 interface IERC20 {
-    function balanceOf(address account) external view returns (uint256);
+    function balanceOf(address account) external view returns (uint);
 }
 
 /// @dev If this is your first time with Forge, read this tutorial in the Foundry Book:
@@ -19,23 +19,23 @@ contract GasKingTest is PRBTest, StdCheats {
     /// @dev A function invoked before each test case is run.
     function setUp() public virtual {
         // Instantiate the contract-under-test.
-//        foo = new GasKing();
+        //        foo = new GasKing();
     }
 
     /// @dev Basic test. Run it with `forge test -vvv` to see the console log.
     function test_Example() external {
         console2.log("Hello World");
-        uint256 x = 42;
-//        assertEq(foo.id(x), x, "value mismatch");
+        uint x = 42;
+        //        assertEq(foo.id(x), x, "value mismatch");
     }
 
     /// @dev Fuzz test that provides random values for an unsigned integer, but which rejects zero as an
     /// input.
     /// If you need more sophisticated input validation, you should use the `bound` utility instead.
     /// See https://twitter.com/PaulRBerg/status/1622558791685242880
-    function testFuzz_Example(uint256 x) external {
+    function testFuzz_Example(uint x) external {
         vm.assume(x != 0); // or x = bound(x, 1, 100)
-//        assertEq(foo.id(x), x, "value mismatch");
+            //        assertEq(foo.id(x), x, "value mismatch");
     }
 
     /// @dev Fork test that runs against an Ethereum Mainnet fork. For this to work, you need to set
@@ -52,8 +52,8 @@ contract GasKingTest is PRBTest, StdCheats {
         vm.createSelectFork({ urlOrAlias: "mainnet", blockNumber: 16_428_000 });
         address usdc = 0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48;
         address holder = 0x7713974908Be4BEd47172370115e8b1219F4A5f0;
-        uint256 actualBalance = IERC20(usdc).balanceOf(holder);
-        uint256 expectedBalance = 196_307_713.810457e6;
+        uint actualBalance = IERC20(usdc).balanceOf(holder);
+        uint expectedBalance = 196_307_713.810457e6;
         assertEq(actualBalance, expectedBalance);
     }
 }
